requirements.txt
numpy
pandas
pillow
tensorflow>=1.3.0
scikit-learn
matplotlib
streamlit
opencv-python
pytorch
torchvision
cython
ninja
gitpython
scipy
scikit-image
h5py
ipython[all]
keras>=2.0.8
imgaug
git+https://github.com/matterport/Mask_RCNN.git